+++ packages/module-realtime/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @graphand/client-module-realtime
 
+## 1.2.3
+
+### Patch Changes
+
+- add support for medias upload in gdx push in cli 💾
+- Updated dependencies
+  - @graphand/client@1.2.6
+
 ## 1.2.2
 
 ### Patch Changes